MarketplaceOrdering updates to tests
