+++ src/SDKs/MarketplaceOrdering/MarketplaceOrdering.Tests/Helpers/MarketplaceOrderingHelper.cs
@@ -26,42 +26,5 @@ namespace MarketplaceOrdering.Tests.Helpers
 
             return null;
         }
-
-        public static string GenerateRandomKey()
-        {
-            byte[] key256 = new byte[32];
-            using (var rngCryptoServiceProvider = RandomNumberGenerator.Create())
-            {
-                rngCryptoServiceProvider.GetBytes(key256);
-            }
-
-            return Convert.ToBase64String(key256);
-        }
-
-        public static string ConvertObjectToJSon<T>(T obj)
-        {
-            return ConvertObjectToJSonAsync(obj);
-        }
-
-        public static string ConvertObjectToJSonAsync(object obj)
-        {
-            if (obj != null)
-            {
-                return (Task.Factory.StartNew(() => JsonConvert.SerializeObject(obj, SerializeMediaTypeFormatterSettings))).Result;
-            }
-            return String.Empty;
-        }
-
-        private static readonly JsonSerializerSettings SerializeMediaTypeFormatterSettings = new JsonSerializerSettings
-        {
-            NullValueHandling = Newtonsoft.Json.NullValueHandling.Include,
-            ContractResolver = new CamelCasePropertyNamesContractResolver(),
-            Converters = new List<JsonConverter>
-            {
-                new StringEnumConverter { CamelCaseText = false },
-            },
-            ConstructorHandling = ConstructorHandling.AllowNonPublicDefaultConstructor
-        };
-
     }
 }
