+++ src/SDKs/MarketplaceOrdering/MarketplaceOrdering.Tests/MarketplaceOrdering.Tests.csproj
@@ -25,4 +25,7 @@
   <ItemGroup>
     <Service Include="{82a7f48d-3b50-4b1e-b82e-3ada8210c358}" />
   </ItemGroup>
+  <ItemGroup>
+    <Folder Include="SessionRecords\" />
+  </ItemGroup>
 </Project>
